+++ eras/alonzo/impl/test/Main.hs
@@ -4,6 +4,7 @@ module Main where
 
 import Cardano.Ledger.Alonzo (Alonzo)
 import Data.Data (Proxy (..))
+import qualified Test.Cardano.Ledger.Alonzo.Binary.CddlSpec as CddlSpec
 import qualified Test.Cardano.Ledger.Alonzo.BinarySpec as BinarySpec
 import Test.Cardano.Ledger.Alonzo.ImpTest ()
 import Test.Cardano.Ledger.Common
@@ -14,4 +15,5 @@ main =
   ledgerTestMain $
     describe "Alonzo" $ do
       BinarySpec.spec
+      CddlSpec.spec
       ImpTestSpec.spec $ Proxy @Alonzo