+++ eras/babbage/impl/test/Main.hs
@@ -4,6 +4,7 @@ module Main where
 
 import Cardano.Ledger.Babbage (Babbage)
 import Data.Data (Proxy (..))
+import qualified Test.Cardano.Ledger.Babbage.Binary.CddlSpec as CddlSpec
 import qualified Test.Cardano.Ledger.Babbage.BinarySpec as BinarySpec
 import Test.Cardano.Ledger.Babbage.ImpTest ()
 import Test.Cardano.Ledger.Common
@@ -14,4 +15,5 @@ main =
   ledgerTestMain $
     describe "Babbage" $ do
       BinarySpec.spec
+      CddlSpec.spec
       ImpTestSpec.spec $ Proxy @Babbage