+++ eras/babbage/test-suite/test/Tests.hs
@@ -7,7 +7,6 @@ import Cardano.Ledger.Babbage (Babbage)
 import Data.Proxy (Proxy (..))
 import System.Environment (lookupEnv)
 import qualified Test.Cardano.Ledger.Babbage.GoldenTranslation as Golden (tests)
-import qualified Test.Cardano.Ledger.Babbage.Serialisation.CDDL as CDDL
 import qualified Test.Cardano.Ledger.Babbage.Serialisation.Tripping as Tripping
 import Test.Cardano.Ledger.Babbage.TxInfo (txInfoTests, txInfoTestsBabbageOnly)
 import Test.Tasty (TestTree, defaultMain, testGroup)
@@ -27,12 +26,10 @@ defaultTests =
     , Tripping.tests
     , txInfoTests (Proxy @Babbage)
     , txInfoTestsBabbageOnly
-    , CDDL.tests 5
     ]
 
 nightlyTests :: TestTree
 nightlyTests =
   testGroup
     "Babbage tests - nightly"
-    [ CDDL.tests 50
-    ]
+    []