+++ eras/conway/test-suite/test/Tests.hs
@@ -6,7 +6,6 @@ import Cardano.Ledger.Conway (Conway)
 import Data.Proxy (Proxy (..))
 import qualified Test.Cardano.Ledger.Babbage.TxInfo as Babbage (txInfoTests)
 import Test.Cardano.Ledger.Conway.GovSnapshot (govSnapshotProps)
-import qualified Test.Cardano.Ledger.Conway.Serialisation.CDDL as CDDL
 import qualified Test.Cardano.Ledger.Conway.TxInfo as Conway (txInfoTests)
 import Test.Tasty (TestTree, defaultMain, testGroup)
 
@@ -17,8 +16,7 @@ defaultTests :: TestTree
 defaultTests =
   testGroup
     "Conway tests"
-    [ CDDL.tests 10
-    , Babbage.txInfoTests (Proxy @Conway)
+    [ Babbage.txInfoTests (Proxy @Conway)
     , Conway.txInfoTests (Proxy @Conway)
     , govSnapshotProps
     ]