+++ eras/mary/impl/test/Main.hs
@@ -5,6 +5,7 @@ module Main where
 import Cardano.Ledger.Mary (Mary)
 import Data.Data (Proxy (..))
 import Test.Cardano.Ledger.Common
+import qualified Test.Cardano.Ledger.Mary.Binary.CddlSpec as CddlSpec
 import qualified Test.Cardano.Ledger.Mary.BinarySpec as BinarySpec
 import Test.Cardano.Ledger.Mary.ImpTest ()
 import qualified Test.Cardano.Ledger.Mary.ValueSpec as ValueSpec
@@ -16,4 +17,5 @@ main =
     describe "Mary" $ do
       ValueSpec.spec
       BinarySpec.spec
+      CddlSpec.spec
       ImpTestSpec.spec $ Proxy @Mary