+++ eras/shelley-ma/test-suite/test/Test/Cardano/Ledger/ShelleyMA/Serialisation.hs
@@ -2,7 +2,6 @@ module Test.Cardano.Ledger.ShelleyMA.Serialisation where
 
 import Test.Cardano.Ledger.Allegra.Translation (allegraEncodeDecodeTests)
 import Test.Cardano.Ledger.Mary.Translation (maryEncodeDecodeTests)
-import Test.Cardano.Ledger.ShelleyMA.Serialisation.CDDL (cddlTests)
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.Golden.Encoding (goldenEncodingTests)
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.Roundtrip (allEraRoundtripTests)
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.Timelocks (timelockTests)
@@ -17,7 +16,6 @@ tests =
     , maryEncodeDecodeTests
     , txBodyTest
     , timelockTests
-    , cddlTests 10
     , goldenEncodingTests
     , allEraRoundtripTests
     ]