+++ eras/conway/test-suite/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-conway-test`
 
+## 1.2.1.2
+
+* Move `cddl-files` to `cardao-ledger-alonzo`
+
 ## 1.2.1.1
 
 * Update CDDL files to reflect the change in header structure in